Updated to fix edge case bug
Whenever all of the electrodes in the model were set to 0 V, the ions would stop registering collisions, and would fly right through walls. I'm making changes to ydot(), evolveIonZDependent(), trajectoryTOFGateDetector(), trajectoryFullLoading(), and the tutorial pdf to fix this. Nothing should change for 99% of use cases, the only change is a slight difference in formatting when working with z-dependent models that use evolveIonZDependent(). Instead of putting zeroes in the symbolic amplitudes arrays for electrodes in different files, you now place NaN's as detailed in section 7.2.2 of the tutorial.